feat: move resourceFilter functions/structs to k8s util
