+++ internal/controller/common/riders/rider.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/numaproj/numaplane/internal/common"
 	"github.com/numaproj/numaplane/internal/controller/config"
-	"github.com/numaproj/numaplane/internal/sync"
 	"github.com/numaproj/numaplane/internal/util/kubernetes"
 	"github.com/numaproj/numaplane/internal/util/logger"
 	apiv1 "github.com/numaproj/numaplane/pkg/apis/numaplane/v1alpha1"
@@ -175,30 +174,30 @@ func GetRidersFromK8S(ctx context.Context, namespace string, ridersList []apiv1.
 	return existingRiders, nil
 }
 
-func VerifyRiders(riders []Rider) error {
+func VerifyRiders(riders []Rider) bool {
 
 	// get global configmap
 	globalConfig, err := config.GetConfigManagerInstance().GetConfig()
 	if err != nil {
-		return err
+		return false
 	}
 
 	// parse permitted Riders and create resource filter
-	riderRules, err := sync.ParseResourceFilter(globalConfig.PermittedRiders)
+	riderRules, err := kubernetes.ParseResourceFilter(globalConfig.PermittedRiders)
 	if err != nil {
-		return err
+		return false
 	}
-	resourcesFilter := &sync.ResourceFilter{
-		IncludedResources: *riderRules,
+	resourcesFilter := &kubernetes.ResourceFilter{
+		IncludedResources: riderRules,
 	}
 
 	// check that each rider is a permitted Group/Kind set in the Numplane controller config
 	for _, rider := range riders {
 		gvk := rider.Definition.GroupVersionKind()
 		if resourcesFilter.IsExcludedResource(gvk.Group, gvk.Kind, "") {
-			return fmt.Errorf("Rider %s of %s/%s is not a permitted Kind", rider.Definition.GetName(), gvk.Group, gvk.Kind)
+			return false
 		}
 	}
 
-	return nil
+	return true
 }
