+++ internal/controller/isbservicerollout/isbservicerollout_controller.go
@@ -1055,9 +1055,9 @@ func (r *ISBServiceRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.
 	}
 
 	// verify that desiredRiders are all permitted Kinds
-	err := riders.VerifyRiders(desiredRiders)
-	if err != nil {
-		return desiredRiders, err
+	// err := riders.VerifyRiders(desiredRiders)
+	if !riders.VerifyRiders(desiredRiders) {
+		return desiredRiders, fmt.Errorf("rider definitions contained unpermitted Kind")
 	}
 
 	return desiredRiders, nil
