+++ internal/controller/pipelinerollout/pipelinerollout_controller.go
@@ -1365,9 +1365,9 @@ func (r *PipelineRolloutReconciler) GetDesiredRiders(rolloutObject ctlrcommon.Ro
 	}
 
 	// verify that desiredRiders are all permitted Kinds
-	err := riders.VerifyRiders(desiredRiders)
-	if err != nil {
-		return desiredRiders, err
+	// err := riders.VerifyRiders(desiredRiders)
+	if !riders.VerifyRiders(desiredRiders) {
+		return desiredRiders, fmt.Errorf("rider definitions contained unpermitted Kind")
 	}
 
 	return desiredRiders, nil
